pyproject.toml
[project]
name = "intellisense-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno==2.3.4",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "babel==2.18.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "blockbuster==1.5.25",
    "boto3==1.42.50",
    "botocore==1.42.50",
    "build==1.4.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.5.0",
    "click==8.3.1",
    "cloudpickle==3.1.2",
    "colorama==0.4.6",
    "courlan==1.3.2",
    "cryptography==44.0.3",
    "dataclasses-json==0.6.7",
    "dateparser==1.3.0",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "durationpy==0.10",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.5.2",
    "fastapi[standard]==0.122.0",
    "fastar==0.8.0",
    "filelock==3.24.2",
    "flatbuffers==25.12.19",
    "forbiddenfruit==0.1.4",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.2.4",
    "groq==0.36.0",
    "grpcio==1.76.0",
    "grpcio-tools==1.75.1",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.2.0",
    "hpack==4.1.0",
    "htmldate==1.9.4",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.4.1",
    "hyperframe==6.1.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jmespath==1.1.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-rs==0.29.1",
    "jsonschema-specifications==2025.9.1",
    "justext==3.0.2",
    "kubernetes==35.0.0",
    "langchain==1.1.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.1.0",
    "langchain-groq==1.1.0",
    "langchain-mcp-adapters==0.1.14",
    "langchain-openai==1.1.0",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.4",
    "langgraph-api==0.5.27",
    "langgraph-checkpoint==3.0.1",
    "langgraph-cli[inmem]==0.4.7",
    "langgraph-prebuilt==1.0.5",
    "langgraph-runtime-inmem==0.19.0",
    "langgraph-sdk==0.2.10",
    "langsmith==0.4.49",
    "lxml==6.0.2",
    "lxml-html-clean==0.4.3",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.1",
    "mcp==1.22.0",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "networkx==3.6.1",
    "numpy==2.3.5",
    "oauthlib==3.3.1",
    "onnxruntime==1.24.1",
    "openai==2.8.1",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-exporter-otlp-proto-http==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "pinecone==8.0.0",
    "pinecone-plugin-assistant==3.0.1",
    "pinecone-plugin-interface==0.0.7",
    "pluggy==1.6.0",
    "posthog==5.4.0",
    "propcache==0.4.1",
    "protobuf==6.33.1",
    "pybase64==1.4.3",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pypdf==6.7.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "pytest==9.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.3",
    "rank-bm25==0.2.2",
    "redis==7.1.0",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rich-toolkit==0.17.0",
    "rignore==0.7.6",
    "rpds-py==0.29.0",
    "s3transfer==0.16.0",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.0",
    "sentence-transformers==5.2.2",
    "sentry-sdk==2.46.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "sse-starlette==2.1.3",
    "starlette==0.41.2",
    "structlog==25.5.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tld==0.13.1",
    "tokenizers==0.22.2",
    "torch==2.10.0",
    "tqdm==4.67.1",
    "trafilatura==2.0.0",
    "transformers==5.1.0",
    "truststore==0.10.4",
    "typer==0.20.0",
    "typer-slim==0.21.2",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "youtube-transcript-api==1.2.4",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
